requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
databricks-sdk==0.102.0
docker==7.1.0
fastapi==0.135.3
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
gunicorn==25.3.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
Mako==1.3.11
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
multidict==6.7.1
numpy==2.4.4
opentelemetry-api==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
packaging==26.1
pandas==2.3.3
pillow==12.2.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
wcwidth==0.6.0
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
